Refactored MemberInfo so it handles no user data
